Create eda.py
Created an exploritory data analysis library for any transformers / classes we think of that will simplify the EDA phase of a Data Science Project.

DFInfo works in it's current form however if improvements are suggested then they can be imprlemented. For example if we were to return the indexes of missing values or something along those lines.